[MASTERDIET-39] Add rest of database models

+++ src/main/java/pl/agh/edu/master_diet/core/model/database/UserPlan.java
@@ -54,7 +54,7 @@ public class UserPlan {
     private Goal goal;
 
     @Column(nullable = false)
-    private Integer currentWeight;
+    private Float currentWeight;
 
     @OneToOne
     @MapsId
@@ -69,5 +69,4 @@ public class UserPlan {
                 .proteins(getProteins())
                 .build();
     }
-
 }
